viejo/Senzu/requirements.txt
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
Django==1.11.1
django-cors-headers==2.0.2
django-filter==1.0.4
django-rest-auth==0.9.1
django-rest-framework-docs==0.1.7
django-rest-swagger==2.1.2
djangorestframework==3.6.3
drf-any-permissions==0.0.1
dry-rest-permissions==0.1.9
itypes==1.1.0
Jinja2==2.9.6
jsonpickle==0.9.4
Markdown==2.6.8
MarkupSafe==1.0
mysql-connector-python==2.0.4
mysqlclient==1.3.10
openapi-codec==1.3.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0

#sudo apt-get install libmysqlclient-dev
#sudo -H pip3 install mysqlclient

